+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 This release adds a new feature to the `Live` client for automatically reconnecting when an unexpected disconnection occurs.
 
 #### Enhancements
+- Added `Reference` data client with `corporate_actions.get_range(...)` method
 - Added `ReconnectPolicy` enumeration
 - Added `reconnect_policy` parameter to the `Live` client to specify client reconnection behavior
 - Added `Live.add_reconnect_callback` method for specifying a callback to handle client reconnections